Fix python thread
Found by: Empus
Patch by: thommey and michaelortmann

Fixes [20:20:20] !!! writing to nonexistent socket: 9 when using putlog() or other functions using sockets not available via thread local storage returned via threaddata() in python thread

the underlying idea is that the dccsockets etc. of the main thread are separated from the one in Tcl threads because that is necessary.
From Python however, it should be (famous last words) safe to access the main thread's socket table, so this PR makes it globally accessible, with only Tcl threads having their own.
The safety requirement is that Python is only called when the main thread isn't operating on them and right now, Python is only called during Tcl binds and in select().