remove colon from accounts, modes (#1786)
Found by: @Robby-, tydel
Patch by: Geo

One-line summary:
Remove colons from trailing IRC commands. Inspircd v3 started putting :'s on all trailing commands, even if they cannot have a space in them (accounts, nicks, etc). This broke how Eggdrop stored account names sent from WHOX, and nicknames contained in MODE commands. And since users blame clients and not servers, now we have to do more work to accommodate.

This is a blunt-force and inelegant patch but it at least fixes the identified issues... there are probably more messages that are needed to be handled but not known at this time, and maybe a larger code revamp could be done. Just not today.

Inspircd claims the following as justification for the change: https://docs.inspircd.org/faq/#why-does-my-client-not-show-mode-changesopped-users-on-joinetc-correctly-when-using-inspircd-v3